+++ lpm/pll_16fft_raw.h
@@ -3,7 +3,7 @@
  *
  * PLL Raw driver for direct PLL manipulation
  *
- * Copyright (C) 2021-2022, Texas Instruments Incorporated
+ * Copyright (C) 2021-2023, Texas Instruments Incorporated
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -35,13 +35,14 @@
  */
 
 #include <types/short_types.h>
+#include <types/sbool.h>
 
 struct pll_raw_data {
 	u32	base;
 	u32	freq_ctrl0;
 	u32	freq_ctrl1;
 	u32	div_ctrl;
-	u16	hsdiv[16];
+	u32	hsdiv[16];
 };
 
 /**
